drivers: qcom: ipc: Initialize the RX SKB header appropriately
The header of the RX Socket buffers(SKBs) is initialized with the maximum
receiveable packet size. When the concerned SKB gets forwarded to another
transport, it leads to passing incorrect packet length and a potential
stall of that IPC Router transport.

Receive the incoming data in the SKB and then initialize the SKB header
with the appropriate packet information.

Change-Id: Ia150e4b3498a481af4a2dacf7d02e9e90a9f935e
Signed-off-by: Karthikeyan Ramasubramanian <kramasub@codeaurora.org>