msm: rndis_ipa: start TX queue immediately
Change the flow control implementation of rndis network interface to
start the TX queue as soon as possible.
When outstanding packets low watermark is hit, call to
netif_wake_queue. This will start the packet transmission in the same
context instead of waiting for network start to schedule
a TX packet transmission.
This change prevents a data stall seen in TCP.
This stall happens as packets are delayed in the network stack
in a busy system. This delay is between starting the TX queue
and the actual transmission of the TX packets by the network stack.

Change-Id: I5de08df89559c9b969e3ed503947ef6b599b0b20
Acked-by: Ady Abraham <adya@qti.qualcomm.com>
Signed-off-by: Ravit Katzav <rkatzav@codeaurora.org>