gpio: qpnp-pin: fix qpnp_pin_probe issues
There are variables which should be declared as u32 instead of int. This
results in misbehavior when there are no gpio/mpp device nodes specified
in the topology. Also, the memory allocation size for q.chip->pmic_pins in
probe is miscalculated.

Hence change the variables to the correct type and update the memory
allocation size calculation for q.chip->pmic_pins.

Change-Id: Ibac3d24a2ad29cf3f605e3aff42a3f3093e40a7d
Signed-off-by: Ke Liu <keliu@codeaurora.org>