usb: gadget: android: Enable RNDIS_QC depending on transport
On some targets the RNDIS gadget function may support BAM-to-BAM
transport, and is handled by the alternate f_qc_rndis driver. In
order to select the correct gadget driver while simplifying
existing userspace composition scripts that already supply "rndis"
as the function name, check the rndis_transports field if
'BAM2BAM_IPA' is supplied, and if so, enable the "rndis_qc"
function instead.

Change-Id: Ia965e1dffd2d74abe07aa6ca3cd9ced4dcba1e0f
Signed-off-by: Jack Pham <jackp@codeaurora.org>