msm: qpnp-power-on: add device tree support to indicate if UVLO happened
Currently, the device will be reboot if UVLO happened. However, we don't
want reboot the device if UVLO happened and we want panic the device in
some situations.

Hence add a qcom,uvlo-panic property to support this feature. If
qcom,uvlo-panic is present, then we will panic the device during power on
sequence. If this property is not present, the device will continue to
boot.

Change-Id: I963e18002f5ce2e9cdfd0ada99ce19b276055164
Signed-off-by: Ke Liu <keliu@codeaurora.org>