usb: dwc3: support LPM during usb bus suspend
This change implements the required sequence for going
in and out of low-power-mode during usb bus suspend.

It is also required to take the core out of LPM in case
ep_queue is called by the upper layers. In this case,
remote wakeup sequence will be initiated once the core
is out of LPM.

CRs-Fixed: 641671
Change-Id: I8076e58db4037502589f04f8cd7e7fd98cdc59f4
Signed-off-by: Bar Weiner <bweiner@codeaurora.org>