Revert "clk: qcom: Check CLKFLAG_INIT_DONE flag for clk_get()"
This reverts commit f5cb564aaf2dc1ebcb2237c1aa37f8c58f54027f.

CLKFLAG_INIT_DONE should be set for all clocks that are ready for use.
It's not correct to acquire and use a clock before this flag is set.
However, there are clock users depend on this wrong behavior on
MSM8916, and it's not trivial to fix.

Revert the change now while a fix for clock users is being
investigated.

Change-Id: I77b3b04c7b597576a4eda2eda5cfe8cab63b5be7
Signed-off-by: Junjie Wu <junjiew@codeaurora.org>