input: synaptics_dsx: correct sysfs permissions
Set correct permissions for write-only sysfs nodes. Show method
of such attributes is set to NULL since the attributes are write-only.
Remove the sysfs entry for the suspend/resume of touch controller
because the suspend/resume user interface is already part of debugfs
entry. Add sysfs entry to read package id of the touch controller.

Change-Id: I22b3980bc415e6749689f9fc0e7345bbd87cb680
Signed-off-by: Sarada Prasanna Garnayak <c_sgarna@codeaurora.org>