usb: gadget: rndis_qc: Enable WCEIS by default
The 'wceis' flag, which is settable in userspace, causes the RNDIS
function to be enumerated with an alternate Class/SubClass/Protocol
corresponding to the Windows "Remote NDIS based Internet Sharing"
device driver. In most cases, this is what we want, so that the
built-in Windows driver will bind to this function without
additional driver installation. Enable this parameter by default;
special cases can explicitly set this to 0.

Change-Id: Ief8b5f26a6b37c1b72bdc029601331d05be49b09
Signed-off-by: Jack Pham <jackp@codeaurora.org>