usb: dwc3: Repurpose enable suspend event property
If suspend events are disabled, the device controller driver
cannot process a SUSPEND event from the host, the gadget
composite driver won't get any notification, and the result
of this is that the peripheral will continue to draw 500mA
over VBUS which is a regression.

The original intent of commit 0f449c05 "usb: dwc3: Add support
to configure Suspend Event" was to add a DT property that decides
whether to allow controller low power mode upon bus suspend, but
it made it so that suspend events are completely disabled when the
flag is not set. Instead, rename this flag and move it to the core
driver; this narrows its scope by using it only to check whether to
enter LPM during peripheral bus suspend. This allows the suspend event
to be re-enabled in all cases so that it can be handled as before.

Change-Id: Ie507eff7e5bc74967e1c9b82f4badb8e895d8043
Signed-off-by: Vamsi Krishna <vskrishn@codeaurora.org>
Signed-off-by: Jack Pham <jackp@codeaurora.org>