usb: dwc3-msm: Fix check for device bus suspend
Rather than checking for mdwc->vbus_active, which can be true
even with a DCP charger connected but without an enumerated
session, check for 'device_bus_suspend' which is true when SDP
or CDP connection is detected. This allows for the controller
to bypass setting of autosuspend and also to power collapse when
either cable is disconnected or when a DCP charger is attached.

Change-Id: I7ecc6e149b52d0b5b0a3ee1d38ad51a0b1c18f97
Signed-off-by: Jack Pham <jackp@codeaurora.org>