msm: vidc: Distinguish between active and passive buses
Introduce support for "passive buses" which are buses aren't tied to the
hardware load.  The buses are just enabled during the power on sequence
to allow for certain clocks (that aren't directly controlled by the
driver) to be forced on.

This also allows us to remove hacks in the code that prevented us from
unvoting for buses when hardware was technically idle.  Given that
passive buses keep their vote until the hardware is powered down, the
hacks within the source code can be replaced by introducing passive bus
clients in the device tree on a target-by-target (or a clock-by-clock)
basis.

Change-Id: Ieac94073f84f2c5d48e96467f69186d0944ad50c
Signed-off-by: Deva Ramasubramanian <dramasub@codeaurora.org>
Signed-off-by: Maheshwar Ajja <majja@codeaurora.org>