power_supply: introduce properties for supporting external SMB chargers
There are situations when one needs to notify involved power supply
nodes that type detection is about to commence and they should
modify their state and hardware to allow it.

For example, the D+/D- lines are shared by USB PHY and the Charger
driver where the phy pullsdown the D+/D- lines. When USB is inserted
the charger power supply node will need to notify the USB power supply
node to release the pulldowns and allow for type detection.

This property serves as that indication.

Also, create a helper function to set this property.

CRs-Fixed: 1086015
Change-Id: Id896d02a19cf8bdef7e520d452aa764a2d48b7b2
Signed-off-by: Abhijeet Dharmapurikar <adharmap@codeaurora.org>
Signed-off-by: Hemant Kumar <hemantk@codeaurora.org>