properly finish that update: Leds: Add op sequece for GPIO Led driver.
Pinctrl is used for gpio control. Remove old gpio
dependency and pinctrl support in driver.
The control logic of different flash driver IC
may be different for flash mode and torch mode.
So adding sequence config in dtsi to make sure
this driver can be used for different driver IC.

Signed-off-by: Kui Wang <kuiw@codeaurora.org>
Change-Id: I5036cd80bd3520faf62d76714471a8bcff7b5421